Renaming overlay_zipper file and internal mod name for consistency
