+++ src/lib.rs
@@ -128,7 +128,7 @@ mod trie_node;
 mod write_zipper;
 mod product_zipper;
 mod prefix_zipper;
-mod overlay;
+mod overlay_zipper;
 mod trie_ref;
 mod dense_byte_node;
 pub(crate) mod line_list_node;
