+++ src/zipper.rs
@@ -17,7 +17,7 @@ pub use crate::write_zipper::*;
 pub use crate::trie_ref::*;
 pub use crate::zipper_head::*;
 pub use crate::product_zipper::{ProductZipper, ProductZipperG};
-pub use crate::overlay::{OverlayZipper};
+pub use crate::overlay_zipper::{OverlayZipper};
 pub use crate::prefix_zipper::{PrefixZipper};
 
 use crate::zipper_tracking::*;
